Move ComponentInterfaceSymbol to separate header file
