+++ src/Shuttle.h
@@ -12,6 +12,7 @@
 #define __AUDACITY_SHUTTLE__
 
 #include "../include/audacity/ComponentInterface.h"
+#include "ComponentInterfaceSymbol.h"
 
 class ComponentInterfaceSymbol;
 class WrappedType;
