+++ src/widgets/NumericTextCtrl.h
@@ -19,6 +19,7 @@
 
 #include <memory>
 #include "../../include/audacity/ComponentInterface.h"
+#include "ComponentInterfaceSymbol.h"
 #include <vector>
 #include <wx/setup.h> // for wxUSE_* macros
 #include <wx/defs.h>
